\name{Service functions}
\alias{NLSLL.3}

\title{
Service functions
}
\description{
The functions documented here are internally used by the package and they are not to be used elsewhere
}
\usage{
NLSLL.3(predictor, b, d, ED50)
}
\arguments{
\item{predictor}{response variable}
\item{b}{shepe parameter}
\item{d}{truncation parameter}
\item{ED50}{location parameter}
}
